+++ src/v8/config.ts
@@ -5,6 +5,7 @@ export const uploadNamspace = "dzapi";
 
 export const apiEndpoint = "https://d5dn8hh4ivlobv6682ep.apigw.yandexcloud.net";
 export const presignPut = apiEndpoint + "/presign-put";
+export const presignDelete = apiEndpoint + "/presign-delete";
 export const tokenGet = apiEndpoint + "/token/get";
 
 export const brCdn = "https://br.cdn.dos.zone";